# spec/edited/openapi.yaml
openapi: 3.0.2
info:
  title: Wellbore DDMS OSDU
  description: build local
  version: '0.2'
paths:
...
  '/ddms/v2/wellbores/{wellboreid}':
    get:
      tags:
        - Wellbore
      summary: 'Get the Wellbore using wks:wellbore:1.0.6 schema'
      description: |-
        Get the Wellbore object using its **id**.  <p>If the wellbore kind is
                *wks:wellbore:1.0.6* returns the record directly</p> <p>If the wellbore
                kind is different *wks:wellbore:1.0.6* it will get the raw record and
                convert the results to match the *wks:wellbore:1.0.6*. If convertion is
                not possible returns an error **500**
      operationId: get_wellbore
      parameters:
        - required: true
          schema:
            title: Wellboreid
            type: string
          name: wellboreid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/wellbore'
        '404':
          description: Wellbore not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
    delete:
      tags:
        - Wellbore
      summary: Delete the wellbore. The API performs a logical deletion of the given record
      operationId: del_wellbore
      parameters:
        - required: true
          schema:
            title: Wellboreid
            type: string
          name: wellboreid
          in: path
        - description: Whether or not to delete records children
          required: false
          schema:
            title: Recursive
            type: boolean
            description: Whether or not to delete records children
            default: false
          name: recursive
          in: query
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '204':
          description: Record deleted successfully
        '404':
          description: Wellbore not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/wellbores/{wellboreid}/versions':
    get:
      tags:
        - Wellbore
      summary: Get all versions of the Wellbore
      operationId: get_wellbore_versions
      parameters:
        - required: true
          schema:
            title: Wellboreid
            type: string
          name: wellboreid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordVersions'
        '404':
          description: Wellbore not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/wellbores/{wellboreid}/versions/{version}':
    get:
      tags:
        - Wellbore
      summary: 'Get the given version of the Wellbore using wks:wellbore:1.0.6 schema'
      description: |-
        "Get the Wellbore object using its **id**.  <p>If the wellbore kind is
                *wks:wellbore:1.0.6* returns the record directly</p> <p>If the wellbore
                kind is different *wks:wellbore:1.0.6* it will get the raw record and
                convert the results to match the *wks:wellbore:1.0.6*. If convertion is
                not possible returns an error **500**
      operationId: get_wellbore_version
      parameters:
        - required: true
          schema:
            title: Wellboreid
            type: string
          name: wellboreid
          in: path
        - required: true
          schema:
            title: Version
            type: integer
          name: version
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/wellbore'
        '404':
          description: Wellbore not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  /ddms/v2/wellbores:
    post:
      tags:
        - Wellbore
      summary: 'Create or update the Wellbores using wks:wellbore:1.0.6 schema'
      operationId: post_wellbore
      parameters:
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      requestBody:
        content:
          application/json:
            schema:
              title: Wellbores
              type: array
              items:
                $ref: '#/components/schemas/wellbore'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUpdateRecordsResponse'
        '400':
          description: Missing mandatory parameter or unknown parameter
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
...
components:
  schemas:
...
    CreateUpdateRecordsResponse:
      title: CreateUpdateRecordsResponse
      type: object
      properties:
        recordCount:
          title: Recordcount
          type: integer
        recordIds:
          title: Recordids
          type: array
          items:
            type: string
        skippedRecordIds:
          title: Skippedrecordids
          type: array
          items:
            type: string
...
    GeoJsonFeatureCollection:
      title: GeoJsonFeatureCollection
      required:
        - features
        - type
      type: object
      properties:
        bbox:
          title: Bbox
          type: array
          items:
            type: number
        features:
          title: Features
          type: array
          items:
            $ref: '#/components/schemas/GeoJsonFeature'
        type:
          $ref: '#/components/schemas/Type_2'
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        errors:
          title: Errors
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    LinkList:
      title: LinkList
      type: object
      properties: {}
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    MetaItem:
      title: MetaItem
      required:
        - kind
        - persistableReference
      type: object
      properties:
        kind:
          title: Reference Kind
          allOf:
            - $ref: '#/components/schemas/Kind'
          description: 'The kind of reference, unit, measurement, CRS or azimuth reference.'
        name:
          title: Name or Symbol
          type: string
          description: The name of the CRS or the symbol/name of the unit
        persistableReference:
          title: Persistable Reference
          type: string
          description: The persistable reference string uniquely identifying the CRS or Unit
        propertyNames:
          title: Attribute Names
          type: array
          items:
            type: string
          description: 'The list of property names, to which this meta data item provides Unit/CRS context to. Data structures, which come in a single frame of reference, can register the property name, others require a full path like "data.structureA.propertyB" to define a unique context.'
        propertyValues:
          title: Attribute Names
          type: array
          items:
            type: string
          description: 'The list of property values, to which this meta data item provides Unit/CRS context to. Typically a unit symbol is a value to a data structure; this symbol is then registered in this propertyValues array and the persistableReference provides the absolute reference.'
        uncertainty:
          title: Uncertainty
          type: number
          description: The uncertainty of the values measured given the unit or CRS unit.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
    PlssLocation:
      title: PlssLocation
      required:
        - range
        - section
        - township
      type: object
      properties:
        aliquotPart:
          title: Aliquot Part
          type: string
          description: 'A terse, hierarchical reference to a piece of land, in which successive subdivisions of some larger area.'
        range:
          title: Range
          type: string
          description: 'Range, also known as Rng, R; a measure of the distance east or west from a referenced principal meridian, in units of six miles.'
        section:
          title: Section Number
          type: integer
          description: Section number (between 1 and 36)
        township:
          title: Township
          type: string
          description: 'Township, also known as T or Twp; (1) Synonym for survey township, i.e., a square parcel of land of 36 square miles, or (2) A measure of the distance north or south from a referenced baseline, in units of six miles'
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    RecordVersions:
      title: RecordVersions
      type: object
      properties:
        recordId:
          title: Recordid
          type: string
        versions:
          title: Versions
          type: array
          items:
            type: integer
...
    Shape:
      title: Shape
      enum:
        - build and hold
        - deviated
        - double kickoff
        - horizontal
        - S-shaped
        - vertical
        - unknown
      description: An enumeration.
    SimpleElevationReference:
      title: SimpleElevationReference
      required:
        - elevationFromMsl
      type: object
      properties:
        elevationFromMsl:
          title: Elevation from MSL
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: 'The elevation above mean sea level (MSL), at which the vertical origin is 0.0. The ''unitKey'' is further defined in ''frameOfReference.units''.'
        name:
          title: Elevation Reference Name
          type: string
          description: The name of the Elevation Reference.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    TagDictionary:
      title: TagDictionary
      type: object
      properties: {}
      description: Used for data model allows extra fields which are not declared initially in the pydantic model
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    ValueWithUnit:
      title: ValueWithUnit
      required:
        - unitKey
        - value
      type: object
      properties:
        unitKey:
          title: Unit Key
          type: string
          description: Unit for value of the corresponding attribute for the domain object in question. The key can be looked up in the 'frameOfReference.units' for further details.
        value:
          title: Value
          type: number
          description: Value of the corresponding attribute for the domain object in question.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    app__model__model_curated__Legal:
      title: Legal
      type: object
      properties:
        legaltags:
          title: Legal Tags
          type: array
          items:
            type: string
          description: 'The list of legal tags, see compliance API.'
        otherRelevantDataCountries:
          title: Other Relevant Data Countries
          type: array
          items:
            type: string
          description: 'The list of other relevant data countries using the ISO 2-letter codes, see compliance API.'
        status:
          title: Legal Status
          type: string
          description: The legal status.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    core_dl_geopoint:
      title: core_dl_geopoint
      required:
        - latitude
        - longitude
      type: object
      properties:
        latitude:
          title: Latitude
          maximum: 90
          minimum: -90
          type: number
          description: 'The latitude value in degrees of arc (dega). Value range [-90, 90].'
        longitude:
          title: Longitude
          maximum: 180
          minimum: -180
          type: number
          description: 'The longitude value in degrees of arc (dega). Value range [-180, 180]'
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    geographicPosition:
      title: geographicPosition
      required:
        - crsKey
        - elevationFromMsl
        - latitude
        - longitude
      type: object
      properties:
        crsKey:
          title: CRS Key
          type: string
          description: 'The ''crsKey'', which can be looked up in the ''frameOfReference.crs'' for further details.'
        elevationFromMsl:
          title: Elevation from MSL
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: 'Elevation from Mean Seal Level, downwards negative. The unit definition is found via ''elevationFromMsl.unitKey'' in ''frameOfReference.units'' dictionary.'
        latitude:
          title: Native Latitude
          type: number
          description: Native or original latitude (unit defined by CRS)
        longitude:
          title: Native Longitude
          type: number
          description: Native or original longitude (unit defined by CRS)
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    wellbore:
      title: wellbore
      type: object
      properties:
        acl:
          title: Access Control List
          allOf:
            - $ref: '#/components/schemas/TagDictionary'
          description: The access control tags associated with this entity.
        ancestry:
          title: Ancestry
          allOf:
            - $ref: '#/components/schemas/LinkList'
          description: 'The links to data, which constitute the inputs.'
        data:
          title: Wellbore Data
          allOf:
            - $ref: '#/components/schemas/wellboreData'
          description: Wellbore data container
        id:
          title: Wellbore ID
          type: string
          description: The unique identifier of the wellbore
        kind:
          title: Wellbore Kind
          type: string
          description: Well-known wellbore kind specification
          default: 'osdu:wks:wellbore:0.0.1'
        legal:
          title: Legal Tags
          allOf:
            - $ref: '#/components/schemas/app__model__model_curated__Legal'
          description: The geological interpretation's legal tags
        meta:
          title: Frame of Reference Meta Data
          type: array
          items:
            $ref: '#/components/schemas/MetaItem'
          description: 'The meta data section linking the ''unitKey'', ''crsKey'' to self-contained definitions (persistableReference)'
        type:
          title: Entity Type
          type: string
          description: 'The reference entity type as declared in common:metadata:entity:*.'
        version:
          title: Entity Version Number
          type: number
          description: The version number of this wellbore; set by the framework.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
    wellboreData:
      title: wellboreData
      type: object
      properties:
        airGap:
          title: Air Gap
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: The gap between water surface and offshore drilling platform.
        block:
          title: Block
          type: string
          description: 'The block name, in which the wellbore is located.'
        country:
          title: Country
          type: string
          description: 'The country, in which the wellbore is located. The country name follows the convention in ISO 3166-1 ''English short country name'', see https://en.wikipedia.org/wiki/ISO_3166-1'
        county:
          title: County
          type: string
          description: 'The county name, in which the wellbore is located.'
        dateCreated:
          title: Creation Date and Time
          type: string
          description: The UTC date time of the entity creation
          format: date-time
        dateModified:
          title: Last Modification Date and Time
          type: string
          description: The UTC date time of the last entity modification
          format: date-time
        drillingDaysTarget:
          title: Target Drilling Days
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Target days for drilling wellbore.
        elevationReference:
          title: Elevation Reference
          allOf:
            - $ref: '#/components/schemas/SimpleElevationReference'
          description: 'The wellbore''s elevation reference from mean sea level (MSL), positive above MSL. This is where MD == 0 and TVD == 0'
        externalIds:
          title: Array of External IDs
          type: array
          items:
            type: string
          description: 'An array of identities (e.g. some kind if URL to be resolved in an external data store), which links to external realizations of the same entity.'
        field:
          title: Field
          type: string
          description: 'The field name, to which the wellbore belongs.'
        formationAtTd:
          title: Formation at TD
          type: string
          description: The name of the formation at the wellbore's total depth.
        formationProjected:
          title: Formation Projected
          type: string
          description: The name of the formation at the wellbore's projected depth. This property is questionable as there is not precise documentation available.
        hasAchievedTotalDepth:
          title: Has Total Depth Been Achieved Flag
          type: boolean
          description: 'True ("true" of "1") indicates that the wellbore has acheieved total depth. That is, drilling has completed. False ("false" or "0") indicates otherwise. Not given indicates that it is not known whether total depth has been reached.'
          default: true
        isActive:
          title: Is Active Flag
          type: boolean
          description: 'True (="1" or "true") indicates that the wellbore is active. False (="0" or "false") indicates otherwise. It is the servers responsibility to set this value based on its available internal data (e.g., what objects are changing).'
        kickOffMd:
          title: Kick-off MD
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: The kick-off point in measured depth (MD); for the main well the kickOffMd is set to 0.
        kickOffTvd:
          title: Kick-off MD
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Kickoff true vertical depth of the wellbore; for the main wellbore the kickOffMd is set to 0.
        locationWGS84:
          title: Wellbore Shape WGS 84
          allOf:
            - $ref: '#/components/schemas/GeoJsonFeatureCollection'
          description: A 2D GeoJSON FeatureCollection defining wellbore location or trajectory in WGS 84 CRS.
        name:
          title: Wellbore Name
          type: string
          description: The wellbore name
        operator:
          title: Operator
          type: string
          description: The operator of the wellbore.
        permitDate:
          title: Permit Date
          type: string
          description: The wellbore's permit date.
          format: date
        permitNumber:
          title: Permit Number
          type: string
          description: The wellbore's permit number or permit ID.
        plssLocation:
          title: US PLSS Location
          allOf:
            - $ref: '#/components/schemas/PlssLocation'
          description: A location described by the Public Land Survey System (United States)
        propertyDictionary:
          title: Property Dictionary
          type: object
          description: 'A dictionary structure, i.e. key/string value pairs, to carry additional wellbore properties.'
        relationships:
          title: Relationships
          allOf:
            - $ref: '#/components/schemas/wellborerelationships'
          description: The related entities.
        shape:
          title: Wellbore Shape
          allOf:
            - $ref: '#/components/schemas/Shape'
          description: POSC wellbore trajectory shape.
        spudDate:
          title: Spud Date
          type: string
          description: 'The date and time when activities to drill the borehole begin to create a hole in the earth. For a sidetrack, this is the date kickoff operations began. The format follows ISO 8601 YYYY-MM-DD extended format'
          format: date
        state:
          title: State
          type: string
          description: 'The state name, in which the wellbore is located.'
        totalDepthMd:
          title: Total MD
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: 'The measured depth of the borehole. If status is plugged, indicates the maximum depth reached before plugging. It is recommended that this value be updated about every 10 minutes by an assigned raw data provider at a site.'
        totalDepthMdDriller:
          title: Total MD Drilled
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: The total depth along the wellbore as reported by the drilling contractor from 'elevationReference'. The unit definition is found via the property's unitKey' in 'frameOfReference.units' dictionary..
        totalDepthMdPlanned:
          title: Total MD Planned
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Planned measured depth for the wellbore total depth.
        totalDepthMdSubSeaPlanned:
          title: Total MD Sub Sea Planned
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Planned measured for the wellbore total depth - with respect to seabed.
        totalDepthProjectedMd:
          title: Total MD Projected
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: The projected total measured depth of the borehole. This property is questionable as there is not precise documentation available.
        totalDepthTvd:
          title: Total TVD
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: 'The true vertical depth of the borehole. If status is plugged, indicates the maximum depth reached before plugging. It is recommended that this value be updated about every 10 minutes by an assigned raw data provider at a site.'
        totalDepthTvdDriller:
          title: Total TVD Drilled
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: 'The total depth true vertical as reported by the drilling contractor from ''elevationReference'', Downwards increasing. The unit definition is found via the property''s unitKey'' in ''frameOfReference.units'' dictionary.'
        totalDepthTvdPlanned:
          title: Total TVD Planned
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Planned true vertical depth for the wellbore total depth.
        totalDepthTvdSubSeaPlanned:
          title: Total TVD Sub Sea Planned
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: Planned true vertical depth for the wellbore total depth - with respect to seabed.
        uwi:
          title: Unique Wellbore Identifier
          type: string
          description: 'The unique wellbore identifier, aka. API number, US well number or UBHI. Codes can have 10, 12 or 14 digits depending on the availability of directional sidetrack (2 digits) and event sequence codes (2 digits).'
        wellHeadElevation:
          title: Well Head Elevation
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: 'The wellbore''s vertical position is an elevation from mean sea level (MSL), positive above MSL.'
        wellHeadGeographic:
          title: 'Well Head Position, Geographic'
          allOf:
            - $ref: '#/components/schemas/geographicPosition'
          description: 'The wellbore''s well head position in the native, geographic CRS; vertical position is an elevation from mean sea level (MSL), positive above MSL.'
        wellHeadProjected:
          title: 'Well Head Position, Projected'
          allOf:
            - $ref: '#/components/schemas/projectedPosition'
          description: 'The wellbore''s well head position in the native, projected CRS; vertical position is an elevation from mean sea level (MSL), positive above MSL.'
        wellHeadWgs84:
          title: WGS 84 Position
          allOf:
            - $ref: '#/components/schemas/core_dl_geopoint'
          description: The wellbore's position in WGS 84 latitude and longitude.
        wellboreNumberGovernment:
          title: Government Number
          type: string
          description: Government assigned wellbore number.
        wellboreNumberOperator:
          title: Operator Number
          type: string
          description: Operator wellbore number.
        wellborePurpose:
          title: Wellbore Purpose
          allOf:
            - $ref: '#/components/schemas/WellborePurpose'
          description: POSC wellbore purpose
        wellboreStatus:
          title: Wellbore Status
          allOf:
            - $ref: '#/components/schemas/WellboreStatus'
          description: POSC wellbore status.
        wellboreType:
          title: Wellbore Type
          allOf:
            - $ref: '#/components/schemas/WellboreType'
          description: Type of wellbore.
      description: Used for data model allows extra fields which are not declared initially in the pydantic model
    wellborerelationships:
      title: wellborerelationships
      type: object
      properties:
        definitiveTimeDepthRelation:
          title: Definitive Time-Depth Relation
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The definitive tome-depth relation providing the MD to seismic travel-time transformation.
        definitiveTrajectory:
          title: Definitive Trajectory
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The definitive trajectory providing the MD to 3D space transformation.
        tieInWellbore:
          title: Tie-in Wellbore
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The tie-in wellbore if this wellbore is a side-track.
        well:
          title: Well
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The well to which this wellbore belongs.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
  securitySchemes:
    OpenDESBearerToken:
      type: http
      scheme: bearer